requirements.txt
asgiref==3.4.1
Django==3.2.5
django-debug-toolbar==3.2.2
djangorestframework==3.12.4
gunicorn==20.1.0
hashids==1.3.1
Pillow==8.3.1
psycopg2-binary==2.8.6
python-decouple==3.1
pytz==2021.1
sqlparse==0.4.1
